refactor: recoil 상태들 atom과 selector로 분류함에 따른 상태 호출 경로 변경
